backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["David Vilela <dvilelaf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
open-autonomy = "^0.14.2"
open-aea-ledger-ethereum = "^1.46.0"
flask = "2.0.2"
petname = "^2.6"
flask-cors = "^4.0.0"
docker = "6.1.2"
schedule = "^1.2.1"
safe-eth-py = "^5.8.0"
clea = "^0.1.0rc4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
